Final Commit
Ready for Review. Added in a youtube video via iframes. Added in two audio elements with sources of mp3 and ogg files. The audio elements contain an embed elements for browsers that don't support source elements. Added in simple styling for the audio elements. Fixed some of the styles for the audio element. Renamed the files to index.html and rnt_audio.css. Added in comments.